% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_spaghetti.R
\name{ggplot_spaghetti}
\alias{ggplot_spaghetti}
\title{Spaghetti Plots using \code{ggplot2}}
\usage{
ggplot_spaghetti(y, id, time, alpha = 0.2, method = "loess", jit = 0,
  group = NULL, wrap = NULL, bw = FALSE, scales = NULL)
}
\arguments{
\item{y}{This is the y-axis parameter to specify. Generally it is a continuous variable.}

\item{id}{This is the id parameter that identifies the unique individuals or units.}

\item{time}{This is the time vector and must be numeric.}

\item{alpha}{Scalar value between [0,1] that specifies the transparencey of the lineplots.}

\item{method}{Character value that specifies which type of method to use for fitting. Optional methods come from stat_smooth() funciton.}

\item{jit}{Scalar value that specifies how much you want to jitter each individual observation. Useful if many of the values share the same y values at a time point.}

\item{group}{Specifies a grouping variable to be used, and will plot it by color on one single plot.}

\item{wrap}{Another possible grouping variable, but this will use `facet_wrap()` to create separate plots for each grouping variable.}

\item{scales}{Specify the scales paremeter for `facet_wrap()` in ggplot2.}
}
\value{
Plots a time series data by each individual/unit with group trends overlayed.
}
\description{
This function allows the user to create spaghetti plots for individuals with time varying covariates. You can also break this down into subgroups to analyze different trentds.
}
\details{
Note that the data must be in long format.
}
\examples{
library(nlme)
data("Orthodont")
attach(Orthodont)
ortho_plot <- ggplot_spaghetti(distance, Subject, age, alpha = 0.3, group = Sex, method = "lm")+
xlab("Age (yrs.)")+
ylab("Distance")+
scale_color_discrete(name="Gender")
ortho_plot

}
